refactor(canvas): colocate Canvas-only hooks under components/Canvas/hooks
The nine hooks introduced for the Canvas split are not reused
elsewhere, so they belong next to the component that owns them
rather than in the shared hooks/ folder. Relative imports inside
each hook adjust accordingly.
