+++ apps/canvas-workspace/src/renderer/src/components/Canvas/hooks/useCanvasMouseHandlers.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef, type MutableRefObject, type RefObject } from 'react';
-import type { CanvasNode } from '../types';
-import type { EdgeInteractionState } from './useEdgeInteraction';
-import type { ResizeEdge } from './useNodeResize';
+import type { CanvasNode } from '../../../types';
+import type { EdgeInteractionState } from '../../../hooks/useEdgeInteraction';
+import type { ResizeEdge } from '../../../hooks/useNodeResize';
 
 interface MarqueeApi {
   active: boolean;
