+++ apps/canvas-workspace/src/renderer/src/components/Canvas/hooks/useCanvasRenderOrder.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
-import type { CanvasNode } from '../types';
-import { computeContainerDepths, isContainerNode } from '../utils/frameHierarchy';
+import type { CanvasNode } from '../../../types';
+import { computeContainerDepths, isContainerNode } from '../../../utils/frameHierarchy';
 
 /**
  * Computes the canvas render order — containers underneath, regular
