+++ apps/canvas-workspace/src/renderer/src/components/Canvas/index.tsx
@@ -12,15 +12,15 @@ import { useCanvasImagePaste } from '../../hooks/useCanvasImagePaste';
 import { useEdgeInteraction } from '../../hooks/useEdgeInteraction';
 import { useShapeDraw } from '../../hooks/useShapeDraw';
 import { useMarqueeSelect } from '../../hooks/useMarqueeSelect';
-import { useCanvasFocusMode } from '../../hooks/useCanvasFocusMode';
-import { useCanvasSelection } from '../../hooks/useCanvasSelection';
-import { useCanvasContextMenu } from '../../hooks/useCanvasContextMenu';
-import { useCanvasNodeActions } from '../../hooks/useCanvasNodeActions';
-import { useCanvasSyncEffects } from '../../hooks/useCanvasSyncEffects';
-import { useCanvasMouseHandlers } from '../../hooks/useCanvasMouseHandlers';
-import { useCanvasPaletteCommands } from '../../hooks/useCanvasPaletteCommands';
-import { useCanvasEdgeHandlers } from '../../hooks/useCanvasEdgeHandlers';
-import { useCanvasRenderOrder } from '../../hooks/useCanvasRenderOrder';
+import { useCanvasFocusMode } from './hooks/useCanvasFocusMode';
+import { useCanvasSelection } from './hooks/useCanvasSelection';
+import { useCanvasContextMenu } from './hooks/useCanvasContextMenu';
+import { useCanvasNodeActions } from './hooks/useCanvasNodeActions';
+import { useCanvasSyncEffects } from './hooks/useCanvasSyncEffects';
+import { useCanvasMouseHandlers } from './hooks/useCanvasMouseHandlers';
+import { useCanvasPaletteCommands } from './hooks/useCanvasPaletteCommands';
+import { useCanvasEdgeHandlers } from './hooks/useCanvasEdgeHandlers';
+import { useCanvasRenderOrder } from './hooks/useCanvasRenderOrder';
 import { useAppShell } from '../AppShellProvider';
 import type { CanvasNode } from '../../types';
 import type { CanvasNodeRenameRequest } from '../../types/ui-interaction';
